com.openai.unity 7.7.1 (#196)
- More Function utilities and invoking methods
  - Added FunctionPropertyAttribute to help better inform the feature how to format the Function json
  - Added FromFunc<,> overloads for convenance
  - Fixed invoke args sometimes being casting to wrong type
  - Added additional protections for static and instanced function calls
  - Added additional tool utilities:
    - Tool.ClearRegisteredTools
    - Tool.IsToolRegistered(Tool) - Tool.TryRegisterTool(Tool)
- Updated Sample Chat Behaviour with updated tool implementations
- com.utilities.rest -> 2.5.3

+++ OpenAI/Packages/com.openai.unity/Runtime/Audio/AudioTranslationRequest.cs
@@ -16,7 +16,7 @@ namespace OpenAI.Audio
         /// The audio file to translate, in one of these formats: flac, mp3, mp4, mpeg, mpga, m4a, ogg, wav, or webm.
         /// </param>
         /// <param name="model">
-        /// ID of the model to use.
+        /// ID of the model to use. Only whisper-1 is currently available.
         /// </param>
         /// <param name="prompt">
         /// Optional, An optional text to guide the model's style or continue a previous audio segment.<br/>
@@ -37,7 +37,7 @@ namespace OpenAI.Audio
             string model = null,
             string prompt = null,
             AudioResponseFormat responseFormat = AudioResponseFormat.Json,
-            int? temperature = null)
+            float? temperature = null)
             : this(File.OpenRead(audioPath), Path.GetFileName(audioPath), model, prompt, responseFormat, temperature)
         {
         }
@@ -136,7 +136,7 @@ namespace OpenAI.Audio
         public string AudioName { get; }
 
         /// <summary>
-        /// ID of the model to use.
+        /// ID of the model to use. Only whisper-1 is currently available.
         /// </summary>
         public string Model { get; }
